finished the function to sort the subject user can choose from when adding his education, split the component in two smaller ones
